Add Open Graph protocol support
See http://ogp.me/ for information on the Open Graph protocol.

This adds a bit of metadata to the webpage. This allows Facebook, Twitter, and other crawlers to build a "preview" for the website. In this case, the title will be "Is Taylor Swift Single?", and the description will say "Yes". We could also add support for a thumbnail image.